Merge as-centered branch, refactor folders structure and added groups

+++ templates/hosts/alpine/nodhcp/provision.sh
@@ -1,5 +1,5 @@
 #!/bin/sh
-# Internal template
+# nodhcp template
 set -e
 if [ -z $MILXCGUARD ] ; then exit 1; fi
 DIR=`dirname $0`
@@ -10,6 +10,8 @@ cd `dirname $0`
 # DEBIAN_FRONTEND=noninteractive apt-get install -y thunderbird
 
 # Disable DHCP and do DNS config
-#sed -i "s/.*dhcp.*//" /etc/network/interfaces
+if [ ! -f /etc/network/keepdhcp ]
+then
 sed -i "s/dhcp/manual/" /etc/network/interfaces
 echo "domain $domain\nsearch $domain\nnameserver $ns" > /etc/resolv.conf
+fi
